requirements.txt
openai-whisper @ git+https://github.com/openai/whisper.git
openai==0.27.8
torch==2.0.1
torchaudio==2.0.2
python-dotenv==1.0.0
numpy<2.0
